change wordHeight type from int to double for improved precision in word positioning

+++ lib/src/game/word_search_game.dart
@@ -35,7 +35,7 @@ class WordSearchGame extends FlameGame with TapCallbacks, DragCallbacks {
   double timeLeft = 180;
   bool isGameStarted = false;
   bool isPaused = false;
-  int wordHeight = 20;
+  double wordHeight = 20;
 
   final List<Color> wordColors = [
     Colors.green,
